requirements.txt
asgiref==3.10.0
Django==5.2
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.9
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
